One dynamic animator component
