+++ docs/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.1.0
+
+### Changed
+
+* Refactored animators into 2 components that use an animationType prop to delegate.
+
 ## 1.0.10
 
 ### Added
