+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-simple-animators",
-  "version": "1.0.10",
+  "version": "1.1.0",
   "description": "A collection of simple, reusable react-native animator wrapper components",
   "main": "index.js",
   "scripts": {
